requirements.txt
# Renata - Meeting Intelligence System Dependencies

# === Web Framework (FastAPI Production Stack) ===
fastapi
uvicorn[standard]
jinja2
python-multipart
itsdangerous            # for session middleware
starlette

# === Core Audio & Transcription ===
faster-whisper
librosa
numpy
soundfile

# === AI & LLM Orchestration ===
google-generativeai
langchain-ollama
ollama
torch
# nemo_toolkit[asr]     # Uncomment if using local NeMo diarization
pandas

# === Browser Automation (Pilot Bot) ===
playwright
playwright-stealth
pyperclip

# === PDF Generation & Reporting ===
reportlab

# === Google API Integrations (Calendar / Gmail / Drive) ===
google-api-python-client
google-auth-oauthlib
google-auth-httplib2

# === RAG / AI Search Pipeline ===
langchain
langchain-community
sentence-transformers
chromadb
pdfplumber

# === Utilities ===
loguru
scikit-learn
python-dotenv
omegaconf
requests